pyproject.toml
[project]
name = "openkylin"
version = "0.1.0"
description = "A lightweight, modular AI Agent framework - 轻量级模块化AI智能体框架"
authors = [
    {name = "OpenKylin Team", email = "team@openkylin.dev"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["ai", "agent", "llm", "openai", "claude", "ollama", "rag", "plugin"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "aiohttp>=3.9.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]
# Vector store backends
chroma = ["chromadb>=0.4.0"]
qdrant = ["qdrant-client>=1.7.0"]
# Model providers
openai = ["openai>=1.10.0"]
anthropic = ["anthropic>=0.18.0"]
ollama = ["ollama>=0.1.0"]

[project.scripts]
openkylin = "openkylin.cli.main:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["openkylin*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = false
